travel crud for admin

+++ TravelAgency/src/main/java/com/example/TravelAgency/repository/TravelRepository.java
@@ -3,5 +3,5 @@ package com.example.TravelAgency.repository;
 import com.example.TravelAgency.model.Travel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TravelRepository extends JpaRepository<Travel, Integer> {
+public interface TravelRepository extends JpaRepository<Travel, Long> {
 }
